fix(Card): fixed Card issue and renamed formal Card component to Hero
