+++ examples/create-react-app/package.json
@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@wfp/icons": "^0.0.57",
-    "@wfp/ui": "^1.2.0-alpha.4",
-    "node-sass": "^4.9.3",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-scripts": "2.0.4"
+    "@wfp/icons": "^0.0.82",
+    "@wfp/ui": "^1.10.2",
+    "node-sass": "^4.14.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-scripts": "4.0.0",
+    "twig": "^1.15.2"
   },
   "scripts": {
     "start": "react-scripts start",
