+++ examples/create-react-app/src/App.js
@@ -1,7 +1,4 @@
-/*
-Direct import via webpack-loader
 import '@wfp/ui/src/globals/scss/styles.scss';
-*/
 
 import React, { Component } from 'react';
 import { Icon } from '@wfp/ui';
