+++ examples/create-react-app/src/App.scss
@@ -1,5 +1,3 @@
-@import '~@wfp/ui/src/globals/scss/styles.scss';
-
 .App {
   text-align: center;
 }
