+++ package.json
@@ -135,6 +135,7 @@
     "@storybook/addon-docs": "6.1.0-alpha.30",
     "@storybook/addon-jest": "6.1.0-alpha.30",
     "@storybook/addon-links": "6.1.0-alpha.30",
+    "@storybook/addon-viewport": "^6.0.28",
     "@storybook/addons": "6.1.0-alpha.30",
     "@storybook/react": "6.1.0-alpha.30",
     "@storybook/theming": "6.1.0-alpha.30",
