--- src/components/CardNew/README.mdx
@@ -1,1 +1,0 @@
-## Usage
