--- src/components/CardNew/_cardnew.scss
@@ -1,133 +1,0 @@
-@import '../../globals/scss/colors';
-@import '../../globals/scss/vars';
-@import '../../globals/scss/layout';
-@import '../../globals/scss/typography';
-@import '../../globals/scss/helper-mixins';
-@import '../../globals/scss/import-once';
-@import '../../globals/scss/css--reset';
-@import '../../globals/scss/text-gradient';
-
-.#{$prefix}--card-box {
-  box-shadow: 0 1px 8px rgba(0, 0, 0, 0.15);
-  // font-size: 12px;
-}
-
-.#{$prefix}--photo-cardnew--text-card {
-  padding: $spacing-md;
-}
-
-.#{$prefix}--photo-cardnew__info__metadata {
-  margin: 0;
-  margin-bottom: $spacing-2xs;
-  @include typescale('omega');
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.#{$prefix}--photo-cardnew__info__subtitle {
-  margin: 0;
-  @include typescale('zeta');
-}
-
-.#{$prefix}--photo-cardnew__info__title {
-  margin: 0;
-  /* @include typescale('zeta');
-  text-transform: uppercase;
-  letter-spacing: 0.05em; */
-  & + .#{$prefix}--photo-cardnew__info__subtitle {
-    margin-top: $spacing-xs;
-  }
-}
-
-.#{$prefix}--photo-cardnew__background {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  width: inherit;
-  background-size: cover;
-  background-position: 50% 50%;
-  z-index: 100;
-  overflow: hidden;
-}
-
-.#{$prefix}--photo-cardnew--image-header {
-  .#{$prefix}--header-photo {
-    width: 100%;
-    height: 200px;
-    object-fit: cover;
-    margin: 0;
-  }
-  .#{$prefix}--photo-cardnew__info {
-    background: var(ui-01);
-    padding: $spacing-md;
-  }
-  .#{$prefix}--photo-cardnew__info__background {
-    display: none;
-  }
-  .#{$prefix}--photo-cardnew__info__title {
-    color: $interactive-01;
-  }
-}
-
-.#{$prefix}--photo-cardnew--overlay {
-  .#{$prefix}--photo-cardnew__info {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: inherit;
-    padding: $spacing-md * 1.1 $spacing-md;
-    z-index: 100;
-    overflow: hidden;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    color: $color__white;
-    background: rgba(#000000, 0.5);
-  }
-  // .#{$prefix}--photo-cardnew__info {
-
-  // }
-  .#{$prefix}--photo-cardnew__info__background {
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    filter: blur(6px);
-    background-size: cover;
-    transform: scale(1.05);
-    z-index: -1;
-    height: rem(200px);
-    @include breakpoint(bp--md--major) {
-      height: rem(260px);
-    }
-    &:after {
-      content: '';
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      transition: all 0.1s;
-      background: #000000;
-      opacity: 0.6;
-    }
-  }
-}
-
-.#{$prefix}--photo-cardnew--overlay {
-  height: rem(300px);
-  @include breakpoint(bp--md--major) {
-    height: rem(500px);
-  }
-  box-shadow: none;
-  .#{$prefix}--photo-cardnew__info {
-    padding: $spacing-lg;
-  }
-  .#{$prefix}--photo-cardnew__info__background {
-    height: rem(300px);
-    @include breakpoint(bp--md--major) {
-      height: rem(500px);
-    }
-  }
-}
