+++ src/components/Hero/Hero.hbs
