+++ src/documentation/Credits/Credits.stories.mdx
@@ -3,7 +3,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 import ReactDOMServer from 'react-dom/server';
 
 <Meta
-  title="Documentation/Credits"
+  title="Getting started/Credits"
   subtitle=""
   parameters={{
     previewTabs: {
