+++ src/globals/scss/styles.scss
@@ -78,8 +78,8 @@ $deprecations--message: 'Deprecated code was found, this code will be removed be
 @import '../../components/FormHint/form-hint.scss';
 @import '../../components/Link/link';
 @import '../../components/Button/button.scss';
+@import '../../components/Hero/hero';
 @import '../../components/Card/card';
-@import '../../components/CardNew/cardnew';
 @import '../../components/ContentSwitcher/content-switcher';
 @import '../../components/List/list';
 @import '../../components/Loading/loading';
