+++ src/index.js
@@ -8,7 +8,7 @@ export Breadcrumb from './components/Breadcrumb';
 export BreadcrumbItem from './components/BreadcrumbItem';
 export BreadcrumbHome from './components/BreadcrumbHome';
 export Button from './components/Button';
-export Card from './components/Card';
+export Hero from './components/Hero';
 export Checkbox from './components/Checkbox';
 export Credits from './components/Credits';
 export ContentSwitcher from './components/ContentSwitcher';
